feat(nuxt): refactor everything to use the latest Nuxt4 stack
