+++ .env.example
@@ -1,0 +1,1 @@
+NUXT_PUBLIC_I18N_BASE_URL=
