+++ .gitignore
@@ -22,3 +22,10 @@ logs
 .env
 .env.*
 !.env.example
+
+# Eslint cache
+.eslintcache
+
+# pnpm store and compile cache
+.pnpm-store
+node-compile-cache
