+++ .husky/pre-commit
@@ -1,0 +1,3 @@
+cd front
+pnpm exec lint-staged
+pnpm lint --fix --cache
