+++ .husky/pre-push
@@ -1,0 +1,3 @@
+cd front
+pnpm test run
+pnpm typecheck
