+++ .lintstagedrc.json
@@ -1,0 +1,3 @@
+{
+	"*.{js,ts,vue,json,css,yml,yaml,md,html}": "prettier --write"
+}
