+++ .npmrc
@@ -1,0 +1,2 @@
+shamefully-hoist=true
+save-prefix=~
