+++ .prettierrc.json
@@ -1,0 +1,5 @@
+{
+	"plugins": ["prettier-plugin-tailwindcss"],
+	"useTabs": true,
+	"htmlWhitespaceSensitivity": "ignore"
+}
