+++ .vscode/settings.json
@@ -1,53 +1,17 @@
 {
-	// We use ESLint for linting, so every time we save a file, it will be linted
-	"editor.codeActionsOnSave": {
-		"source.fixAll.eslint": "explicit"
+	"files.associations": {
+		"*.css": "tailwindcss"
 	},
-
-	// We use Prettier for all languages it supports and format the code on save for everyone to have the same code style
-	// Why not multiple language-specific settings ? Because it doesn't override user language-specific settings (https://github.com/microsoft/vscode-docs/issues/5796)
-	"[javascript]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
+	"editor.quickSuggestions": {
+		"strings": "on"
 	},
-	"[vue]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[typescript]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[css]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[html]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[json]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[markdown]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-	"[yaml]": {
-		"editor.defaultFormatter": "esbenp.prettier-vscode",
-		"editor.formatOnSave": true
-	},
-
-	// We use Tabs to indent our code so everyone can choose their own tab size without affecting the project
-	"editor.autoIndent": "full",
-	"editor.insertSpaces": false,
-	"editor.detectIndentation": false,
-
-	// We use Unix line endings to avoid conflicts between Windows and Unix systems
-	"files.eol": "\n",
-	"editor.stickyTabStops": false,
-
-	// Can't turn off extensions recommendations so user is told to install them every time they open the project, this is done to enforce code formatting and linting that is done by these extensions
-	"extensions.ignoreRecommendations": false,
+	"tailwindCSS.classAttributes": ["class", "ui"],
+	"tailwindCSS.experimental.classRegex": [
+		["ui:\\s*{([^)]*)\\s*}", "(?:'|\"|`)([^']*)(?:'|\"|`)"]
+	],
+	"i18n-ally.sourceLanguage": "fr",
+	"i18n-ally.displayLanguage": "fr",
+	"i18n-ally.localesPaths": ["i18n/locales"],
+	"i18n-ally.keystyle": "nested",
+	"i18n-ally.enabledFrameworks": ["vue"]
 }
