+++ app/assets/css/tailwind.css
@@ -1,0 +1,23 @@
+@source "../../../nuxt.config.ts";
+
+@font-face {
+	font-family: "HankenGrotesk";
+	font-style: normal;
+	font-weight: 100 900;
+	font-display: swap;
+	src: url("~/assets/fonts/HankenGrotesk.woff2") format("woff2");
+}
+
+@font-face {
+	font-family: "HankenGrotesk";
+	font-style: italic;
+	font-weight: 100 900;
+	font-display: swap;
+	src: url("~/assets/fonts/HankenGrotesk-Italic.woff2") format("woff2");
+}
+
+@theme {
+	--font-sans:
+		"HankenGrotesk", ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
+		"Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+}
