+++ app/assets/data/projects.json
@@ -1,0 +1,36 @@
+[
+	{
+		"date": "2023-2024",
+		"title": "Mon Guichet - nouvelle version",
+		"description": "La nouvelle version front du produit \"Mon Guichet\" est en cours de développement au sein de Logitud. Elle utilise notre librairie MG_UI pour ses composants principaux.",
+		"img": "mg.webp",
+		"link": "https://mulhouse.staging.numesia.fr",
+		"technos": ["Nuxt", "HeadlessUI", "TailwindCSS", "Flowbite"]
+	},
+	{
+		"date": "2023-2024",
+		"title": "MG_UI - Librairie de composants Nuxt",
+		"description": "Utilisée au sein de Logitud, cette librairie comporte des composants dynamiques, customisables et accessibles. Elle est accessible sur NPM et sera bientôt open-source sous license MIT et disponible sur GitHub.",
+		"img": "mg_ui.webp",
+		"link": "https://www.npmjs.com/package/@numesia/mg_ui",
+		"technos": ["Nuxt", "HeadlessUI", "TailwindCSS", "Flowbite"]
+	},
+	{
+		"date": "2024",
+		"title": "Site personnel",
+		"description": "Quoi de mieux qu'un site personnel pour se présenter et exposer ses projets et compétences ? Développé avec des technos que j'affectionne, il est open-source sous license MIT.",
+		"img": "personal-website.webp",
+		"link": "https://nicolaswalter.fr",
+		"github": "https://github.com/wanicolas/Personal-Website",
+		"technos": ["Nuxt", "TailwindCSS"]
+	},
+	{
+		"date": "2024",
+		"title": "Bullet Ballet",
+		"description": "Jeu platformer de tir en 2D en cours de développement avec Phaser. Le but des deux joueurs est d'éliminer l'autre avec des armes et inclut la possibilité de récupérer des objets.",
+		"img": "bullet-ballet.webp",
+		"link": "https://github.com/wanicolas/Bullet-Ballet",
+		"github": "https://github.com/wanicolas/Bullet-Ballet",
+		"technos": ["Phaser", "Vite", "TypeScript"]
+	}
+]
