--- assets/css/main.css
@@ -1,15 +1,0 @@
-@font-face {
-  font-family: "HankenGrotesk";
-  font-style: normal;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("~/assets/fonts/HankenGrotesk.woff2") format("woff2");
-}
-
-@font-face {
-  font-family: "HankenGrotesk";
-  font-style: italic;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("~/assets/fonts/HankenGrotesk-Italic.woff2") format("woff2");
-}
