+++ i18n/i18n.config.ts
@@ -1,0 +1,29 @@
+export default defineI18nConfig(() => {
+	return {
+		datetimeFormats: {
+			fr: {
+				short: {
+					year: "numeric",
+					month: "short",
+					day: "numeric",
+				},
+				numericDateTime: {
+					year: "numeric",
+					month: "numeric",
+					day: "numeric",
+					hour: "numeric",
+					minute: "numeric",
+				},
+			},
+		},
+		numberFormats: {
+			fr: {
+				currency: {
+					style: "currency",
+					currency: "EUR",
+					notation: "standard",
+				},
+			},
+		},
+	};
+});
