+++ i18n/locales/fr.json
@@ -1,0 +1,1 @@
+{}
