+++ package.json
@@ -1,26 +1,42 @@
 {
-  "name": "nuxt-app",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.9.0",
-    "@nuxtjs/color-mode": "^3.3.2",
-    "@nuxtjs/tailwindcss": "^6.11.2",
-    "eslint": "^9.9.0",
-    "eslint-plugin-vue": "^9.27.0",
-    "globals": "^15.9.0",
-    "nuxt": "^4.0.3",
-    "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.6",
-    "typescript-eslint": "^8.2.0",
-    "vue": "^3.4.14",
-    "vue-router": "^4.2.5"
-  }
+	"name": "nuxt-app",
+	"type": "module",
+	"private": true,
+	"scripts": {
+		"build": "nuxt build",
+		"dev": "nuxt dev",
+		"generate": "nuxt generate",
+		"lint": "eslint",
+		"preview": "nuxt preview",
+		"preinstall": "npx only-allow pnpm",
+		"postinstall": "nuxt prepare",
+		"prepare": "cd .. && husky front/.husky || true",
+		"test": "vitest",
+		"test run": "vitest run",
+		"typecheck": "nuxt typecheck"
+	},
+	"dependencies": {
+		"@nuxtjs/color-mode": "3.5.2",
+		"@nuxtjs/i18n": "10.0.6",
+		"nuxt": "^4.0.3",
+		"vue": "^3.5.20",
+		"vue-router": "^4.5.1"
+	},
+	"devDependencies": {
+		"@nuxt/eslint": "1.9.0",
+		"@nuxt/test-utils": "3.19.2",
+		"@nuxtjs/tailwindcss": "~6.14.0",
+		"@vue/test-utils": "~2.4.6",
+		"eslint": "~9.34.0",
+		"eslint-config-prettier": "~10.1.8",
+		"happy-dom": "~18.0.1",
+		"husky": "~9.1.7",
+		"lint-staged": "~16.1.6",
+		"playwright-core": "~1.55.0",
+		"prettier": "3.6.2",
+		"prettier-plugin-tailwindcss": "~0.6.14",
+		"typescript": "^5.9.2",
+		"vitest": "~3.2.4",
+		"vue-tsc": "~3.0.7"
+	}
 }
