--- prettier.config.js
@@ -1,7 +1,0 @@
-export default {
-  plugins: ["prettier-plugin-tailwindcss"],
-  rules: {
-    useTabs: true,
-  },
-  tailwindConfig: "tailwind.config.js",
-};
