--- tailwind.config.js
@@ -1,21 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-const defaultTheme = require("tailwindcss/defaultTheme");
-
-export default {
-  content: [
-    "./components/**/*.{js,vue,ts}",
-    "./layouts/**/*.vue",
-    "./pages/**/*.vue",
-    "./plugins/**/*.{js,ts}",
-    "./app.vue",
-    "./error.vue",
-  ],
-  theme: {
-    extend: {
-      fontFamily: {
-        sans: ['"HankenGrotesk"', ...defaultTheme.fontFamily.sans],
-      },
-    },
-  },
-  darkMode: "class",
-};
